Rust build servwer
Former-commit-id: 1ac02ff0b1780e5e8da582806f41124efa22222f
Former-commit-id: 7fdb5eed8a7fe8f37e5596d05d9601c036f921ef
Former-commit-id: f1a6f97b5ff0f5a2c7a0f5c2cfb0edb44984b85d

+++ worker/worker.py
@@ -85,7 +85,8 @@ def executeCompileTask(user, backend):
 		print(str(errors))
 		sendEmail("Halite Bot Compilation Error", "<h2>The bot that you recently submitted to the Halite competition would not compile on our servers.</h2> <p>Our autocompile script <b>thought that your bot was written in \""+language+".\"</b> If that is incorrect, please change your code's file extensions to <code>cpp</code> and <code>h</code> for C++11, <code>java</code> for Java 7, and <code>py</code> for Python3. Please make sure that your <b>main file is named MyBot</b> (not main, not BasicBot).</p> <b>Here is a description of the compilation error</b>:<br><pre><code>"+"<br>".join(errors)+"</code></pre>", user["email"])
 	backend.compileResult(int(user["userID"]), didCompile, language)
-	shutil.rmtree(workingPath)
+	if os.path.isdir(workingPath):
+		shutil.rmtree(workingPath)
 
 def downloadUsers(users):
 	for user in users:
